+++ styles/templates/repo/index.ts
@@ -1,0 +1,2 @@
+import "./view_content";
+import "./view_list";