+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ##### Template File
 
+- Synchronized repository code list page layout and styles
+- Synchronized repository add file button menu styles
+
 ### 🐞 Fix
 
 - Fixed extra lines under heatmap.